refactor - using one package instead of three
Previously I was having a dedicated package for flag, delete and
utility. This caused quiet a bit of unnecessary complexity. Decided
to bring all internal logic together in one package since seperate
they don't make sense anyway.
Also changed the internal folder to pkg. this is because I can reuse
the logic in in some other applciation.
Moved the example into examples dir and added that to dockerignore so
that we don't accidentally get that in prod docker images.